Hello
Testing

+++ src/gyro.h
@@ -17,13 +17,13 @@ const float SAMPLE_TIME = 0.05f; // 50ms sampling rate
 
 typedef struct {
     int16_t raw_x, raw_y, raw_z;
-    int32_t avg_x, avg_y, avg_z;
+    int16_t avg_x, avg_y, avg_z;
 } GestureData;
 
 typedef struct {
-    int avg_x[SAMPLE];
-    int avg_y[SAMPLE];
-    int avg_z[SAMPLE];
+    int16_t avg_x[SAMPLE];
+    int16_t avg_y[SAMPLE];
+    int16_t avg_z[SAMPLE];
 } Gesture;
 
 #define CTRL_REG1 0x20
@@ -41,7 +41,7 @@ typedef struct {
 #define OUT_X_L 0x28
 
 int window_index = 0;
-int32_t window_x[WINDOW_SIZE], window_y[WINDOW_SIZE], window_z[WINDOW_SIZE] = {0};
+int16_t window_x[WINDOW_SIZE], window_y[WINDOW_SIZE], window_z[WINDOW_SIZE] = {0};
 
 extern GestureData gd={0}; 
 extern Gesture gd_saved = {0};
